Filter drawer component
